requirements.txt
pymupdf==1.24.12
google-generativeai==0.3.0  # Example version
faiss-cpu==1.7.2
sentence-transformers==3.2.1
python-dotenv==0.20.0
huggingface-hub==0.26.1
numpy==1.22.0
